requirements.txt
apache-airflow
tweepy
pandas
logging